feat: make 'margin' style generic
According to the [spec](https://www.w3.org/TR/CSS2/box.html#propdef-margin)
margin can apply to all elements. This makes margin to be applicable to all
elements, and does so through a type erasure. This is okay because when we are
applying styles that are generally applicable, then other general styles can be applied
to the AnyElement. The root type however will be preserved in that it is passed on to
to the "element" field of the html node.