feat(dsl): expand the capabilities of the html dsl
With this change, the HTML eDSL now supports
for loops, arrays, and if statements!

This will be leveraged to simplify the creation of new
notes in that they can now be stored in an array as
metadata then looped over to list them for the site.
Furthermore, this will unlock and simply having individual
pages for each note to make sharing easier.