feat: add new types for use when processing notes
NoteDecodingError will be used to bubble up errors when
preparing all the notes for display (html)

NoteMetadata is the full type for the note that is passed around.
This contains all the pertinent information related to a note such as
the code snippet.

NoteTag is an enum that represents different tags for the notes like
target audience