feat: upgrade the site generation!
Notes are now a first class citizen! They
are generated from the folder structure.

Adding new notes is greatly simplified:
To add a new note, simply add
metadata and a snippet to the appropriate folder
(named with the date as yyyy-mm-dd). That's it!

A list of the improvements
- Improved output when generating the site
- Notes are gathered up front. This means that they
can be used in other paces in addition to the home page
that has all the notes.
- Error handling is much better now. Things are better bubbled
up if something goes wrong.
- Dates for the notes are no longer strings. They are based off
the folder names and converted to and from Swift.